feat: ✨ MongoDB自动建表/索引切换为异步后台服务
- 新增三大后台服务类，异步自动创建索引、时序集合、Capped 集合，支持取消令牌，日志与异常处理更完善
- 所有 UseCreateMongoXxx 扩展方法废弃，统一为 AddMongoXxxCreation，推荐在服务注册阶段调用
- IndexManager 全面异步化，底层全部切换为 MongoDB 驱动异步 API
- 文档与示例同步更新，强调“启动后异步执行，不阻塞应用启动”
- 结构更清晰，职责分离，易维护，建议所有用户升级
- 影响：需将 app.UseCreateMongoXxx() 替换为 builder.Services.AddMongoXxxCreation()
